change to use bfloat16

+++ llmfoundry/models/layers/attention.py
@@ -104,7 +104,7 @@ def scaled_multihead_dot_product_attention(
     if is_causal and (not q.size(2) == 1):
         s = max(s_q, s_k)
         if pjrt.using_pjrt():
-            causal_mask = attn_weight.new_ones(s, s, dtype=torch.float32)
+            causal_mask = attn_weight.new_ones(s, s, dtype=torch.bfloat16)
         else:
             causal_mask = attn_weight.new_ones(s, s, dtype=torch.float16)
         causal_mask = causal_mask.tril()
